requirements.txt
praw
python-dotenv
openai
faiss-cpu
numpy
reportlab